# Conflicts:
#	ClientHandler/src/main/java/ir/sharif/gamein2021/ClientHandler/controller/GameDataController.java
#	ClientHandler/src/main/java/ir/sharif/gamein2021/ClientHandler/controller/MainController.java
#	Core/src/main/java/ir/sharif/gamein2021/core/service/AuctionService.java
#	Core/src/main/java/ir/sharif/gamein2021/core/util/RequestTypeConstant.java
#	Core/src/main/java/ir/sharif/gamein2021/core/util/ResponseTypeConstant.java

+++ ClientHandler/src/main/java/ir/sharif/gamein2021/ClientHandler/domain/RFQ/EditNegotiationCostPerUnitRequest.java
@@ -11,5 +11,5 @@ import java.io.Serializable;
 public class EditNegotiationCostPerUnitRequest extends RequestObject implements Serializable
 {
     private Integer negotiationId;
-    private Integer newCostPerUnit;
+    private Float newCostPerUnit;
 }
